PyUnicode_Join():  Missed a spot where I intended a cast from size_t to int.  I sure wish MS would gripe about that!  Whatever, note that the statement above it guarantees that the cast loses no info.

+++ Objects/unicodeobject.c
@@ -4047,7 +4047,7 @@ PyUnicode_Join(PyObject *separator, PyObject *seq)
     sz = reslen + 100;  /* breathing room */
     if (sz < reslen || sz > INT_MAX) /* overflow -- no breathing room */
     	sz = reslen;
-    res = _PyUnicode_New(sz);
+    res = _PyUnicode_New((int)sz);
     if (res == NULL) {
         Py_DECREF(item);
         goto onError;
